examples/gitlab-nodejs/090-unittesting/python/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.10
multidict==4.7.6
python-engineio==3.13.1
python-socketio==4.6.0
six==1.15.0
websockets==8.1
yarl==1.4.2
